Fix namespace and resources after move to other project
