+++ src/SonarScanner.MSBuild.Shim/IgnoredIssues/SarifIssueRemover04.cs
@@ -22,7 +22,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
 
-namespace SonarScanner.MSBuild.IgnoredIssues
+namespace SonarScanner.MSBuild.Shim.IgnoredIssues
 {
     public static class SarifIssueRemover04
     {
