+++ src/SonarScanner.MSBuild.Tasks/Resources.resx
@@ -199,12 +199,6 @@ Error: {1}</value>
   <data name="AnalyzerSettings_ExternalIssueNotSupported" xml:space="preserve">
     <value>External issues are not supported on this version of {0}. SQv7.4+ is required.</value>
   </data>
-  <data name="AnalyzerSettings_ImportAllSettingValue" xml:space="preserve">
-    <value>{0}={1}</value>
-  </data>
-  <data name="AnalyzerSettings_InvalidValueForImportAll" xml:space="preserve">
-    <value>Invalid value for '{0}'. Expecting 'true' or 'false'. Actual: '{1}'. External issues will not be imported.</value>
-  </data>
   <data name="AnalyzerSettings_LanguageNotSpecified" xml:space="preserve">
     <value>Analysis language is not specified</value>
   </data>
