+++ src/SonarScanner.MSBuild/Resources.resx
@@ -117,12 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="AnalyzerSettings_ImportAllSettingValue" xml:space="preserve">
-    <value>{0}={1}</value>
-  </data>
-  <data name="AnalyzerSettings_InvalidValueForImportAll" xml:space="preserve">
-    <value>Invalid value for '{0}'. Expecting 'true' or 'false'. Actual: '{1}'. External issues will not be imported.</value>
-  </data>
   <data name="AssemblyDescription" xml:space="preserve">
     <value>SonarScanner for MSBuild</value>
   </data>
